Yelp based search wired up for iOS and Android -- using OAuth 1.0a clients, hitting API and parsing results. OAuth performed on iOS with Obj-C code from Yelp github, and using scribe-1.3.5 on Android. OAuth client does not go through ASI/cURL so requires implementation of WiFi gating at the search service level. DeCarta is also preserved on a single boolean in AppHost.
Added iOS and Android Yelp POI visualisation, pull down and draw images for Yelp POI results, add versioning for MyPinModel & SearchResultModel serialization.

Also, add GeoNames search query, factory, response parser and dto model. Hit GeoNames service as part of the YelpSearchService query. The YelpSearchService is now querying to create an aggregate of the Yelp results and the GeoNames place result, to augment the Yelp POI with place-searching functionality.

Because the Yelp requests go through a bespoke web request mechanism that implements OAuth 1a, they are not filtered by the application implementation of the IWebLoadRequestFactory, through which disabling downloading over non-WiFi connections is implemented. We must directly perform this filtering in the YelpSearchService by querying the INetworkCapabilities interface.

Add mapping of Yelp categories based on taxonomy downloaded from the Yelp API github. Generate a mapping from nodes in the hierarchy we tag as being on of our categories, and work backword via this mapping to map from a search result leaf category to one of our application categories. It seems that the web API contains some categories which are not present in the taxonomy document so sometimes this approach fails. An SO post here http://stackoverflow.com/a/16069505 suggests the taxonomy document is out of data and that the only solution is to parse the Yelp developer site HTML.

Redesign search result POI detail card to better use data from Yelp and apply review feedback. MyPins are now sorted alphabetically to make easier for user to read, and have search specific icons so can easily be distinguished from the user-created MyPins. POI detail card refactored to provide a separate view implementation for each vendor (currently Yelp, GeoNames & deCarta). Yelp POI card layout significantly change to accommodate review feedback. Reviews & human-readable categories now reside on SearchResultModel -- JSON serialization version incremented to account for this.

Refactor representation of metadata associated with a MyPin -- search results and user created MyPin metadata now hydrated into object model such that events can be dispathed, and pin metadata can be processed. This is done because Yelp pins must be actively processed for refreshing. Should also make things like dispatching web requests slightly simpler.

Refactor my pins to enable callback on pin becoming visible or invisible. Wire this callback up to refresh the search result if it is a Yelp vendor result and sufficient time has passed since last refresh. This is on becoming visible as must be initiated from user action to comply with Yelp license.

Buddy = Malc B.

+++ android/jni/SearchResultMenu/View/SearchMenuView.h
@@ -21,7 +21,7 @@ namespace ExampleApp
                 SearchMenuView(AndroidNativeState& nativeState,
                                const std::string& viewClassName);
 
-                void SetHeader(const std::string& header, bool queryPending, int numResult);
+                void SetHeader(const std::string& header, bool queryPending, size_t numResult);
 
                 void HandleSearchClosed();
