+++ src/SearchResultMenu/View/ISearchResultMenuView.h
@@ -17,7 +17,7 @@ namespace ExampleApp
             public:
                 virtual ~ISearchResultMenuView() { };
 
-                virtual void SetHeader(const std::string& header, bool queryPending, int numResult) = 0;
+                virtual void SetHeader(const std::string& header, bool queryPending, size_t numResult) = 0;
 
                 virtual void InsertSearchClosed(Eegeo::Helpers::ICallback0& callback) = 0;
                 virtual void RemoveSearchClosed(Eegeo::Helpers::ICallback0& callback) = 0;
